fix : 좋아요기능 수정

+++ src/main/java/ktb/week4/userpostlike/UserPostLikeService.java
@@ -1,6 +1,7 @@
 package ktb.week4.userpostlike;
 
 import ktb.week4.post.Post;
+import ktb.week4.post.PostRepository;
 import ktb.week4.post.PostService;
 import ktb.week4.post.postView.PostViewService;
 import ktb.week4.user.User;
@@ -13,27 +14,25 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class UserPostLikeService {
     private final UserPostLikeRepository userPostLikeRepository;
-    private final PostService postService;
-    private final UserService userService;
     private final PostViewService postViewService;
+    private final PostRepository postRepository;
 
     @Transactional
-    public void createUserPostLike(Long postId, User user) {
-        Post post = postService.getPostById(postId);
+    public void addLike(Long postId, User user) {
+        Post post = getPostById(postId);
+
         validateOwner(user, post.getUser());
 
-        UserPostLike userPostLike = UserPostLike.builder()
-                .user(user)
-                .post(post)
-                .build();
-        userPostLikeRepository.save(userPostLike);
-
+        createUserPostLike(user, post);
         postViewService.updateLikeCount(postId, false);
     }
 
+
     @Transactional
-    public void deleteUserPostLike(Long postId, User user) {
-        Post post = postService.getPostById(postId);
+    public void removeLike(Long postId, User user) {
+        Post post = getPostById(postId);
+
+        validateOwner(user, post.getUser());
 
         userPostLikeRepository.deleteByUserAndPost(user, post);
         postViewService.updateLikeCount(postId, true);
@@ -43,5 +42,18 @@ public class UserPostLikeService {
         if (user.getId().equals(postUser.getId())) {
             throw new IllegalArgumentException("post owner can't like own posts");
         }
+    }
+
+    private void createUserPostLike(User user, Post post) {
+        UserPostLike userPostLike = UserPostLike.builder()
+                .user(user)
+                .post(post)
+                .build();
+        userPostLikeRepository.save(userPostLike);
+    }
+
+    private Post getPostById(Long postId) {
+        return postRepository.findById(postId)
+                .orElseThrow(()-> new IllegalArgumentException("post not found"));
     }
 }
